bugfix(validate): also include the 'from' when validating via's/ & viaNot exceptions to cycle rules
fixes the issue raised in #604 - which was introduced in version 11.7.0
