+++ src/validate/matchers.js
@@ -91,19 +91,13 @@ function toDependencyTypesNot(pRule, pDependency) {
   );
 }
 
-const removeLast = (_via, pIndex, pArray) => pIndex !== pArray.length - 1;
-
 function toVia(pRule, pDependency, pGroups) {
   return Boolean(
     !pRule.to.via ||
       (pDependency.cycle &&
-        pDependency.cycle
-          // the last in the cycle is always the module itself, which, for
-          // via & viaNot checks isn't very useful
-          .filter(removeLast)
-          .some((pVia) =>
-            pVia.match(replaceGroupPlaceholders(pRule.to.via, pGroups))
-          ))
+        pDependency.cycle.some((pVia) =>
+          pVia.match(replaceGroupPlaceholders(pRule.to.via, pGroups))
+        ))
   );
 }
 
@@ -111,13 +105,9 @@ function toViaOnly(pRule, pDependency, pGroups) {
   return Boolean(
     !pRule.to.viaOnly ||
       (pDependency.cycle &&
-        pDependency.cycle
-          // the last in the cycle is always the module itself, which, for
-          // via & viaNot checks isn't very useful
-          .filter(removeLast)
-          .every((pVia) =>
-            pVia.match(replaceGroupPlaceholders(pRule.to.viaOnly, pGroups))
-          ))
+        pDependency.cycle.every((pVia) =>
+          pVia.match(replaceGroupPlaceholders(pRule.to.viaOnly, pGroups))
+        ))
   );
 }
 
@@ -125,13 +115,9 @@ function toViaNot(pRule, pDependency, pGroups) {
   return Boolean(
     !pRule.to.viaNot ||
       (pDependency.cycle &&
-        !pDependency.cycle
-          // the last in the cycle is always the module itself, which, for
-          // via & viaNot checks isn't very useful
-          .filter(removeLast)
-          .some((pVia) =>
-            pVia.match(replaceGroupPlaceholders(pRule.to.viaNot, pGroups))
-          ))
+        !pDependency.cycle.some((pVia) =>
+          pVia.match(replaceGroupPlaceholders(pRule.to.viaNot, pGroups))
+        ))
   );
 }
 
@@ -139,13 +125,9 @@ function toviaSomeNot(pRule, pDependency, pGroups) {
   return Boolean(
     !pRule.to.viaSomeNot ||
       (pDependency.cycle &&
-        !pDependency.cycle
-          // the last in the cycle is always the module itself, which, for
-          // via & viaNot checks isn't very useful
-          .filter(removeLast)
-          .every((pVia) =>
-            pVia.match(replaceGroupPlaceholders(pRule.to.viaSomeNot, pGroups))
-          ))
+        !pDependency.cycle.every((pVia) =>
+          pVia.match(replaceGroupPlaceholders(pRule.to.viaSomeNot, pGroups))
+        ))
   );
 }
